refactor: rename factories & business logic
