+++ packages/platform-sdk-btc/source/config.ts
