--- packages/platform-sdk-http/source/helpers.ts
@@ -1,9 +1,0 @@
-export const ensureTrailingSlash = (url: string): string => {
-	const lastCharacter = url.substr(-1);
-
-	if (lastCharacter != "/") {
-		url = url + "/";
-	}
-
-	return url;
-};
