+++ packages/platform-sdk-http/source/request.ts
@@ -1,5 +1,4 @@
 import { HttpClient, HttpResponse } from "./contracts";
-import { ensureTrailingSlash } from "./helpers";
 import { RequestOptions } from "./request.models";
 
 export abstract class AbstractRequest implements HttpClient {
@@ -106,3 +105,13 @@ export abstract class AbstractRequest implements HttpClient {
 
 	protected abstract send(method: string, url: string, data?: { query?: object; data?: any }): Promise<HttpResponse>;
 }
+
+const ensureTrailingSlash = (url: string): string => {
+	const lastCharacter = url.substr(-1);
+
+	if (lastCharacter != "/") {
+		url = url + "/";
+	}
+
+	return url;
+};
