+++ packages/platform-sdk-sol/source/public-key.service.ts
@@ -1,7 +1,7 @@
 import { Coins, Exceptions, IoC, Services } from "@arkecosystem/platform-sdk";
 import { BIP39 } from "@arkecosystem/platform-sdk-crypto";
 
-import { derivePrivateKey, derivePublicKey } from "./helpers";
+import { derivePrivateKey, derivePublicKey } from "./keys";
 
 @IoC.injectable()
 export class PublicKeyService extends Services.AbstractPublicKeyService {
